Fixed zero on hand inventory records deletion

+++ springboots/Microservices/mssc-beer-inventory/src/main/java/elearning/sfg/beer/inventory/services/AllocationServiceImpl.java
@@ -99,6 +99,10 @@ public class AllocationServiceImpl implements AllocationService {
                 beerOrderLine.setQuantityAllocated(allocatedQty + inventory);
                 beerInventory.setQuantityOnHand(0);
 
+                //beerInventoryRepository.delete(beerInventory);
+            }
+
+            if (beerInventory.getQuantityOnHand() == 0) {
                 beerInventoryRepository.delete(beerInventory);
             }
         });
